refactor: 순위 counting 을 RankCounter 에게 맡기고 LottoList 에서 foreach 를 제공한다.
