+++ src/main/java/lotto/domain/entity/LottoList.java
@@ -1,7 +1,6 @@
 package lotto.domain.entity;
 
 import lotto.domain.Lotto;
-import lotto.domain.RankCounter;
 import lotto.domain.generator.AutomaticLottoNumbersGenerator;
 
 import java.util.ArrayList;
@@ -30,36 +29,8 @@ public final class LottoList {
         return lottoList.get(index);
     }
 
-    public RankCounter compareWith(Lotto winningLotto, Number bonusNumber) {
-        RankCounter rankCounter = new RankCounter();
-        boolean matchBonus = false;
-        for (Lotto purchased : lottoList) {
-            matchBonus = isMatchBonus(purchased, bonusNumber);
-            checkRank(purchased.confirmWinning(winningLotto), rankCounter, matchBonus);
-        }
-        return rankCounter;
-    }
-
-    private boolean isMatchBonus(Lotto purchased, Number bonusNumber) {
-        return purchased.isNumber(bonusNumber);
-    }
-
-    private void checkRank(int count, RankCounter rankCounter, boolean matchBonus) {
-        if (count == 3) {
-            rankCounter.addCount(Rank.FIFTH);
-        }
-        if (count == 4) {
-            rankCounter.addCount(Rank.FOURTH);
-        }
-        if (count == 5 && !matchBonus) {
-            rankCounter.addCount(Rank.THIRD);
-        }
-        if (count == 5 && matchBonus) {
-            rankCounter.addCount(Rank.SECOND);
-        }
-        if (count == 6) {
-            rankCounter.addCount(Rank.FIRST);
-        }
+    public Iterable<? extends Lotto> foreach() {
+        return lottoList;
     }
 
     @Override
